+++ hazelcast/include/hazelcast/client/map/DeleteRequest.h
@@ -15,20 +15,18 @@ namespace hazelcast {
         namespace map {
             class HAZELCAST_API DeleteRequest : public impl::PortableRequest {
             public:
-                DeleteRequest(const std::string& name, serialization::Data& key, int threadId);
+                DeleteRequest(const std::string &name, serialization::Data &key, long threadId);
 
                 int getFactoryId() const;
 
                 int getClassId() const;
 
-                void write(serialization::PortableWriter& writer) const;
-
-
+                void write(serialization::PortableWriter &writer) const;
 
             private:
-                serialization::Data& key;
+                serialization::Data &key;
                 std::string name;
-                int threadId;
+                long threadId;
             };
         }
     }
