+++ hazelcast/include/hazelcast/client/map/PutTransientRequest.h
@@ -15,21 +15,20 @@ namespace hazelcast {
         namespace map {
             class HAZELCAST_API PutTransientRequest : public impl::PortableRequest {
             public:
-                PutTransientRequest(const std::string& name, serialization::Data& key, serialization::Data& value, int threadId, long ttl);
+                PutTransientRequest(const std::string &name, serialization::Data &key, serialization::Data &value, long threadId, long ttl);
 
                 int getFactoryId() const;
 
                 int getClassId() const;
 
-                void write(serialization::PortableWriter& writer) const;
-
+                void write(serialization::PortableWriter &writer) const;
 
 
             private:
-                serialization::Data& key;
-                serialization::Data& value;
+                serialization::Data &key;
+                serialization::Data &value;
                 std::string name;
-                int threadId;
+                long threadId;
                 long ttl;
             };
         }
