+++ hazelcast/include/hazelcast/client/map/TryRemoveRequest.h
@@ -15,20 +15,19 @@ namespace hazelcast {
         namespace map {
             class HAZELCAST_API TryRemoveRequest : public impl::PortableRequest {
             public:
-                TryRemoveRequest(const std::string& name, serialization::Data& key, int threadId, long timeout);
+                TryRemoveRequest(const std::string &name, serialization::Data &key, long threadId, long timeout);
 
                 int getFactoryId() const;
 
                 int getClassId() const;
 
-                void write(serialization::PortableWriter& writer) const;
-
+                void write(serialization::PortableWriter &writer) const;
 
 
             private:
-                serialization::Data& key;
+                serialization::Data &key;
                 std::string name;
-                int threadId;
+                long threadId;
                 int timeout;
             };
         }
