+++ hazelcast/src/hazelcast/client/map/PutRequest.cpp
@@ -10,7 +10,7 @@
 namespace hazelcast {
     namespace client {
         namespace map {
-            PutRequest::PutRequest(const std::string &name, serialization::Data &key, serialization::Data &value, int threadId, long ttl)
+            PutRequest::PutRequest(const std::string &name, serialization::Data &key, serialization::Data &value, long threadId, long ttl)
             :name(name)
             , key(key)
             , value(value)
@@ -30,7 +30,7 @@ namespace hazelcast {
 
             void PutRequest::write(serialization::PortableWriter &writer) const {
                 writer.writeUTF("n", name);
-                writer.writeInt("t", threadId);
+                writer.writeLong("t", threadId);
                 writer.writeLong("ttl", ttl);
                 serialization::ObjectDataOutput &out = writer.getRawDataOutput();
                 key.writeData(out);
