+++ hazelcast/src/hazelcast/client/map/SetRequest.cpp
@@ -11,7 +11,7 @@
 namespace hazelcast {
     namespace client {
         namespace map {
-            SetRequest::SetRequest(const std::string& name, serialization::Data& key, serialization::Data& value, int threadId, long ttl)
+            SetRequest::SetRequest(const std::string &name, serialization::Data &key, serialization::Data &value, long threadId, long ttl)
             :name(name)
             , key(key)
             , value(value)
@@ -28,11 +28,11 @@ namespace hazelcast {
                 return PortableHook::SET;
             }
 
-            void SetRequest::write(serialization::PortableWriter& writer) const {
+            void SetRequest::write(serialization::PortableWriter &writer) const {
                 writer.writeUTF("n", name);
-                writer.writeInt("t", threadId);
+                writer.writeLong("t", threadId);
                 writer.writeLong("ttl", ttl);
-                serialization::ObjectDataOutput& out = writer.getRawDataOutput();
+                serialization::ObjectDataOutput &out = writer.getRawDataOutput();
                 key.writeData(out);
                 value.writeData(out);
             };
